refactor(cli): Move the conformance suites out of main.rs
main.rs claimed to keep "only the Cli/Commands types and main" while 1,447 of
its 1,785 lines were a test module. Four unrelated suites lived there, sharing
nothing but Cli::command(), and every failure landed under the same `tests::`
prefix. Splitting them by subject makes the claim true again and gives a
failure a name that says which contract broke.

- move the suites to src/conformance/{mod,flags,capabilities,docs}.rs; test
  names, assertions and registries are unchanged
- keep the shared clap-tree helpers (walk, sub, arg, assert_long) and the
  derive check in conformance/mod.rs, where all three suites reach them
- split docs_cli_examples_never_install_into_a_gated_service into its two
  subjects, so a gated service missing from the references stops failing under
  the name of the examples check
- add a cli_smoke row pinning that `capabilities` under a nonexistent
  `--config` still exits 0 and prints the name list, which is the D7 promise
  no in-process test can witness
- repoint the comments in args.rs and commands/capabilities.rs at the files
  the two tests now live in
